qps/hat: resize two attributes in 'qhat_tree_enumerator_t'
In order to save some room for later, reduce the size of the attributes
'pos' and 'count' to 16 bits: they are relative to position and entry
count in a QHAT leaf and 16 bits will always be sufficient for this
usage, as we can see when looking at the QHAT constants for each value
sizes:

VALUE_SIZE  LEAF_INDEX_BITS  LEAVES_PER_COMPACT  LEAVES_PER_FLAT
8           12               4096                1636
16          11               2048                1363
32          10               1024                1023
64          10               1024                1023
128         10               1024                1023

Purpose: make some room for new attributes we want to add in a
subsequent patch.

Priv-Id: a5373304ade09d86c18e2e8dac4da8a8fe694fa8

+++ src/qps-hat.h
@@ -608,13 +608,13 @@ typedef struct qhat_tree_enumerator_t {
      * If the current leaf is a "flat", it's only the index of the value as
      * there is no key array.
      */
-    uint32_t pos;
+    uint16_t pos;
 
     /* Only when the local array in memory is a "compact".
      * Number of elements in the compact.
      * May be unsynchronized with "memory.compact->count" if elements where
      * added or removed in the qhat. */
-    uint32_t count;
+    uint16_t count;
 
     /* Current leaf of the enumerator, cached here for quicker access. */
     qhat_node_const_memory_t memory;
